Move generate-tutorials into view/ folder.
