+++ client/src/view/generate-tutorials/components/weekday-slots/IconForTab.tsx
